Fix win64 compilation error
R=jbauman@chromium.org
TBR=jbauman@chromium.org

Review URL: https://codereview.chromium.org/61743014

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@234039 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 1461bbdafb1268fc789ed3fc02415b3fe09668c8

+++ gpu/command_buffer/client/mapped_memory.h
@@ -33,7 +33,7 @@ class GPU_EXPORT MemoryChunk {
 
   // Gets the size of the chunk.
   unsigned int GetSize() const {
-    return shm_.size;
+    return static_cast<unsigned int>(shm_.size);
   }
 
   // The shared memory id for this chunk.
